Rename Mandelbrot demo to Fractals and update docs
