+++ index.html
@@ -13,7 +13,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <p id="page-info">You are viewing the animath page. Try <a href="#/mandelbrot">#/mandelbrot</a> for the Mandelbrot demo.</p>
+    <p id="page-info">You are viewing the animath page. Try <a href="#/fractals">#/fractals</a> for the fractals demo.</p>
     <textarea id="html-source" readonly style="width:100%;height:200px;"></textarea>
     <script>
       const srcBox = document.getElementById('html-source');
