+++ src/animations/Fractals/README.md
@@ -1,0 +1,3 @@
+# Fractals 2D
+
+This basic component renders several 2D fractals to a canvas. Choose between the Mandelbrot set, a Julia set and the Burning Ship or Multibrot variants.
