--- src/animations/Mandelbrot/README.md
@@ -1,3 +1,0 @@
-# Mandelbrot 2D
-
-This basic component renders the Mandelbrot set to a 2D canvas on mount. It is intended as a starting point for future extensions such as interactive zooming or 3D visualisation.
