+++ src/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import Mandelbrot2D from './animations/Mandelbrot/Mandelbrot2D';
+import Fractals2D from './animations/Fractals/Fractals2D';
 
 const routes: Record<string, JSX.Element> = {
   '/': <App />,
-  '/mandelbrot': <Mandelbrot2D />
+  '/fractals': <Fractals2D />
 };
 
 function getRoute(): JSX.Element {
